requirements.txt
# core
fastapi==0.115.12
pydantic-settings==2.9.1
jinja2==3.1.6
uvicorn==0.34.2
itsdangerous==2.1.2
structlog==23.1.0
python-dotenv==1.0.0
aiohttp[speedups]==3.8.5
aiofiles==23.1.0
moviepy==1.0.3
pyzmq==25.1.0
pillow==9.5.0

segno==1.5.2
millify==0.1.1
isodate==0.6.1
emoji-country-flag==1.3.2

# databse
tinyflux==0.4.0
tinydb==4.8.0

# 3rd parties
yt-dlp==2025.4.30
google-api-python-client==2.91.0
google-auth-oauthlib==1.0.0
google-auth==2.21.0
google-auth-httplib2==0.1.0

instagrapi==2.1.3

